# fix to speed up charpoly for large sparse matrices